Properly name JS/CSS enrty points
