+++ source/javascripts/scripts.js
@@ -1,0 +1,3 @@
+var shareThis = require("./modules/sharethis.js")
+
+shareThis()
