+++ source/layouts/layout.slim
@@ -16,8 +16,9 @@ html
     = feed_tag :atom, "#{config[:content_url]}/feed.xml", title: "#{config[:content_title]} – All articles"
     - categories.each do |category|
       = feed_tag :atom, "#{config[:content_url]}/#{category.slug}.xml", title: "#{config[:content_title]} – #{category.title} articles"
-    = stylesheet_link_tag :default
+    = stylesheet_link_tag :styles
     = javascript_include_tag "//platform-api.sharethis.com/js/sharethis.js#property=#{ENV["SHARETHIS"]}&product=inline-share-buttons"
+    = javascript_include_tag :scripts
 
   body
